+++ android/app/src/main/java/com/aadarshlokhande/productive/widget/ProductivePlusWidget.java
@@ -1,0 +1,6 @@
+package com.aadarshlokhande.productive.widget;
+
+import com.reactnativeandroidwidget.RNWidgetProvider;
+
+public class ProductivePlusWidget extends RNWidgetProvider {
+}